requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
sqlparse==0.4.3
whitenoise==6.2.0
